Remove link annotation from the error message string
This removes the annotation [link] and [/link] from the code that is
supposed to clean the JSON schema error messages. However, while these
annotations might improve the readability of the URL to the JSON schema,
they are wrongly displayed on GitHub when running amdirt validate as
part of the GitHub workflow.